requirements.txt
Flask
python-dotenv
firebase-admin
requests
requests-oauthlib
flask-login
flask-moment
Werkzeug